requirements.txt
Flask==3.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.1.3
